requirements.txt
absl-py
altair
asttokens
attrs
blinker
boto3
botocore
cachetools
certifi
chardet
charset-normalizer
click
colorama
contourpy
cycler
decorator
executing
filelock
fire
Flask
fonttools
fsspec
gitdb
GitPython
grpcio
huggingface-hub
idna
intel-openmp
ipython
itsdangerous
jedi
Jinja2
jmespath
jsonschema
jsonschema-specifications
kiwisolver
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mkl
mpmath
networkx
numpy
opencv-python
opencv-python-headless
packaging
pandas
parso
pillow
prompt-toolkit
protobuf
psutil
pure-eval
py-cpuinfo
pyarrow
pybboxes
pydeck
Pygments
pyparsing
python-dateutil
python-dotenv
python-magic
pytz
PyYAML
referencing
requests
requests-toolbelt
rich
roboflow
rpds-py
s3transfer
sahi
scipy
seaborn
setuptools
shapely
six
smmap
stack-data
streamlit
sympy
tbb
tenacity
tensorboard
tensorboard-data-server
termcolor
terminaltables
thop
toml
toolz
torch
torchvision
tornado
tqdm
traitlets
typing_extensions
tzdata
ultralytics
urllib3
watchdog
wcwidth
Werkzeug
wget
wheel
yolov5
